+++ pkg/storage/stores/shipper/bloomshipper/store.go
@@ -30,6 +30,10 @@ type Store interface {
 	ResolveMetas(ctx context.Context, params MetaSearchParams) ([][]MetaRef, []*Fetcher, error)
 	FetchMetas(ctx context.Context, params MetaSearchParams) ([]Meta, error)
 	FetchBlocks(ctx context.Context, refs []BlockRef, opts ...FetchOption) ([]*CloseableBlockQuerier, error)
+	TenantFilesForInterval(
+		ctx context.Context, interval Interval,
+		filter func(tenant string, object client.StorageObject) bool,
+	) (map[string][]client.StorageObject, error)
 	Fetcher(ts model.Time) (*Fetcher, error)
 	Client(ts model.Time) (Client, error)
 	Stop()
@@ -41,7 +45,7 @@ type StoreWithMetrics interface {
 }
 
 type bloomStoreConfig struct {
-	workingDir       string
+	workingDirs      []string
 	numWorkers       int
 	maxBloomPageSize int
 }
@@ -133,6 +137,86 @@ func (b *bloomStoreEntry) FetchBlocks(ctx context.Context, refs []BlockRef, opts
 	return b.fetcher.FetchBlocks(ctx, refs, opts...)
 }
 
+func (b *bloomStoreEntry) TenantFilesForInterval(
+	ctx context.Context,
+	interval Interval,
+	filter func(tenant string, object client.StorageObject) bool,
+) (map[string][]client.StorageObject, error) {
+	tables := tablesForRange(b.cfg, interval)
+	if len(tables) == 0 {
+		return nil, nil
+	}
+
+	tenants := make(map[string][]client.StorageObject, 100)
+	for _, table := range tables {
+		prefix := path.Join(rootFolder, table)
+		level.Debug(b.fetcher.logger).Log(
+			"msg", "listing tenants",
+			"store", b.cfg.From,
+			"table", table,
+			"prefix", prefix,
+		)
+		objects, _, err := b.objectClient.List(ctx, prefix, "")
+		if err != nil {
+			if b.objectClient.IsObjectNotFoundErr(err) {
+				continue
+			}
+
+			return nil, fmt.Errorf("error listing tenants under prefix [%s]: %w", prefix, err)
+		}
+		if len(objects) == 0 {
+			continue
+		}
+
+		// Sort objects by the key to ensure keys are sorted by tenant.
+		cmpObj := func(a, b client.StorageObject) int {
+			if a.Key < b.Key {
+				return -1
+			}
+			if a.Key > b.Key {
+				return 1
+			}
+			return 0
+		}
+		if !slices.IsSortedFunc(objects, cmpObj) {
+			slices.SortFunc(objects, cmpObj)
+		}
+
+		for i := 0; i < len(objects); i++ {
+			tenant, err := b.TenantPrefix(key(objects[i].Key))
+			if err != nil {
+				return nil, fmt.Errorf("error parsing tenant key [%s]: %w", objects[i].Key, err)
+			}
+
+			// Search next object with different tenant
+			var j int
+			for j = i + 1; j < len(objects); j++ {
+				nextTenant, err := b.TenantPrefix(key(objects[j].Key))
+				if err != nil {
+					return nil, fmt.Errorf("error parsing tenant key [%s]: %w", objects[i].Key, err)
+				}
+				if nextTenant != tenant {
+					break
+				}
+			}
+
+			if _, ok := tenants[tenant]; !ok {
+				tenants[tenant] = nil // Initialize tenant with empty slice
+			}
+
+			if filter != nil && !filter(tenant, objects[i]) {
+				continue
+			}
+
+			// Add all objects for this tenant
+			tenants[tenant] = append(tenants[tenant], objects[i:j]...)
+			i = j - 1 // -1 because the loop will increment i by 1
+		}
+	}
+
+	return tenants, nil
+}
+
 // Fetcher implements Store.
 func (b *bloomStoreEntry) Fetcher(_ model.Time) (*Fetcher, error) {
 	return b.fetcher, nil
@@ -193,13 +277,15 @@ func NewBloomStore(
 
 	// TODO(chaudum): Remove wrapper
 	cfg := bloomStoreConfig{
-		workingDir:       storageConfig.BloomShipperConfig.WorkingDirectory,
-		numWorkers:       storageConfig.BloomShipperConfig.BlocksDownloadingQueue.WorkersCount,
+		workingDirs:      storageConfig.BloomShipperConfig.WorkingDirectory,
+		numWorkers:       storageConfig.BloomShipperConfig.DownloadParallelism,
 		maxBloomPageSize: int(storageConfig.BloomShipperConfig.MaxQueryPageSize),
 	}
 
-	if err := util.EnsureDirectory(cfg.workingDir); err != nil {
-		return nil, errors.Wrapf(err, "failed to create working directory for bloom store: '%s'", cfg.workingDir)
+	for _, wd := range cfg.workingDirs {
+		if err := util.EnsureDirectory(wd); err != nil {
+			return nil, errors.Wrapf(err, "failed to create working directory for bloom store: '%s'", wd)
+		}
 	}
 
 	for _, periodicConfig := range periodicConfigs {
@@ -265,6 +351,34 @@ func (b *BloomStore) Block(ref BlockRef) (loc Location) {
 	}
 
 	return
+}
+
+func (b *BloomStore) TenantFilesForInterval(
+	ctx context.Context,
+	interval Interval,
+	filter func(tenant string, object client.StorageObject) bool,
+) (map[string][]client.StorageObject, error) {
+	var allTenants map[string][]client.StorageObject
+
+	err := b.forStores(ctx, interval, func(innerCtx context.Context, interval Interval, store Store) error {
+		tenants, err := store.TenantFilesForInterval(innerCtx, interval, filter)
+		if err != nil {
+			return err
+		}
+
+		if allTenants == nil {
+			allTenants = tenants
+			return nil
+		}
+
+		for tenant, files := range tenants {
+			allTenants[tenant] = append(allTenants[tenant], files...)
+		}
+
+		return nil
+	})
+
+	return allTenants, err
 }
 
 // Fetcher implements Store.